requirements.txt
# Core async libraries
asyncio-mqtt==0.16.1
aiohttp==3.9.1
aioboto3==12.3.0
aiofiles==23.2.1

# AWS SDK
boto3==1.34.34
botocore==1.34.34

# Database drivers
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.1

# Firebase
firebase-admin==6.4.0
google-cloud-firestore==2.14.0

# Excel generation
openpyxl==3.1.2
xlsxwriter==3.1.9

# Data processing
pandas==2.1.4
numpy==1.26.2

# HTTP client
httpx==0.26.0
requests==2.31.0

# Configuration and environment
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Logging and monitoring
structlog==23.2.0
colorlog==6.8.0

# Date and time handling
python-dateutil==2.8.2
pytz==2023.3

# Email functionality
smtplib2==0.2.1
email-validator==2.1.0

# Utilities
typing-extensions==4.9.0
dataclasses-json==0.6.3